Update DiversityController.py
This displays the bar graph dynamically depending on what is selected in the dropbox.
However, what is displayed is hard to understand because when a single job is selected, it displays a very very small, single bar on one side of the graph. 
This makes it difficult to read the content of the bar graph.